Move Numerics and Operators out of AST.scala.

+++ src/main/scala/com/ndsmith3/parakeet/lexer/Lexer.scala
@@ -43,7 +43,7 @@ object Lexer {
   private def parseFloat(str: String, floatString: String): (Option[FloatToken], String) = {
     @tailrec
     def scan(currStr: String, currFloatString: String = floatString): (Option[FloatToken], String) =
-      if (isCompleteNumber(currStr)) (Some(FloatToken(currFloatString.toFloat)), currStr)
+      if (isCompleteNumber(currStr)) (Some(FloatToken(currFloatString.toDouble)), currStr)
       else scan(currStr.tail, currFloatString + currStr.head)
 
     scan(str)
